# openapi.json
{"openapi":"3.1.0","info":{"title":"MovieLand","description":"REST API for managing movie catalog.","version":"1.0.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Auth","description":"Authentication and token management"},{"name":"Movie Rating","description":"Operations for managing user ratings on movies"},{"name":"Genre","description":"Operations related to movie genres"},{"name":"Review","description":"Operations related to movie reviews"},{"name":"Movies","description":"Operations for managing the movie catalog: search, retrieval, creation, and updates."}],"paths":{"/api/v1/movies/{movieId}/rate":{"get":{"tags":["Movie Rating"],"summary":"Get user's rating for a movie","description":"Returns the current user's rating for a specific movie. Requires USER role.","operationId":"findByMovieId","parameters":[{"name":"movieId","in":"path","description":"Movie ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"userDetails","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtUserDetails"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieRatingResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]},"put":{"tags":["Movie Rating"],"summary":"Update a movie rating","description":"Updates the current user's rating for a specific movie. Requires USER role.","operationId":"update","parameters":[{"name":"movieId","in":"path","description":"Movie ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"userDetails","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtUserDetails"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieRatingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieRatingResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]},"post":{"tags":["Movie Rating"],"summary":"Rate a movie","description":"Creates a new rating for a specific movie. Requires USER role.","operationId":"create","parameters":[{"name":"movieId","in":"path","description":"Movie ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"userDetails","in":"query","required":true,"schema":{"$ref":"#/components/schemas/JwtUserDetails"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieRatingRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieRatingResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}},"/api/v1/movies/{id}":{"get":{"tags":["Movies"],"summary":"Get movie by ID","description":"Returns detailed information about a movie by its ID, including price in the selected currency.","operationId":"findById","parameters":[{"name":"id","in":"path","description":"Movie identifier","required":true,"schema":{"type":"integer","format":"int64"},"example":1},{"name":"currency","in":"query","required":false,"schema":{"type":"string","default":"UAH","enum":["UAH","USD","EUR"]}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieDetailResponse"}}}},"500":{"description":"Server error"}}},"put":{"tags":["Movies"],"summary":"Update a movie","description":"Update movie details. Requires ADMIN role.","operationId":"update_1","parameters":[{"name":"id","in":"path","description":"Movie ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateMovieRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieDetailResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]},"delete":{"tags":["Movies"],"summary":"Mark a movie as deleted","description":"Soft-deletes a movie by marking it as deleted. Requires ADMIN role.","operationId":"mark","parameters":[{"name":"id","in":"path","description":"Movie ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"204":{"description":"No Content"},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}},"/api/v1/reviews":{"post":{"tags":["Review"],"summary":"Create a review","description":"Add a review for a specific movie. Requires USER role.","operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReviewResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}},"/api/v1/refresh":{"post":{"tags":["Auth"],"summary":"Refresh token","description":"Issue a new JWT access token using the current valid token","operationId":"refresh","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}},"/api/v1/movies":{"get":{"tags":["Movies"],"summary":"Get all movies","description":"Returns a list of all available movies with filtering and sorting options.","operationId":"findAll","parameters":[{"name":"request","in":"query","required":true,"schema":{"$ref":"#/components/schemas/FindAllMovieRequest"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieResponse"}}}}},"500":{"description":"Server error"}}},"post":{"tags":["Movies"],"summary":"Create a movie","description":"Adds a new movie to the database. Requires ADMIN role.","operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMovieRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MovieDetailResponse"}}}},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}},"/api/v1/movies/{id}/unmark":{"post":{"tags":["Movies"],"summary":"Restore a marked movie","description":"Restores a previously soft-deleted movie. Requires ADMIN role.","operationId":"unmark","parameters":[{"name":"id","in":"path","description":"Movie ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK"},"500":{"description":"Server error"},"401":{"description":"No auth"},"403":{"description":"Access denied"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}},"/api/v1/login":{"post":{"tags":["Auth"],"summary":"Login","description":"Authenticate user by email and password, returns JWT access token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"500":{"description":"Server error"}}}},"/api/v1/movies/random":{"get":{"tags":["Movies"],"summary":"Get random movies","description":"Returns a list of 3 random movies from the catalog","operationId":"random","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieResponse"}}}}},"500":{"description":"Server error"}}}},"/api/v1/movies/genre/{genreId}":{"get":{"tags":["Movies"],"summary":"Get movies by genre","description":"Returns all movies belonging to the specified genre","operationId":"findByGenreId","parameters":[{"name":"genreId","in":"path","description":"Genre ID","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MovieResponse"}}}}},"500":{"description":"Server error"}}}},"/api/v1/genres":{"get":{"tags":["Genre"],"summary":"Get all genres","description":"Returns a list of all available genres","operationId":"findAll_1","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/GenreResponse"}}}}},"500":{"description":"Server error"}}}},"/api/v1/logout":{"delete":{"tags":["Auth"],"summary":"Logout","description":"Blacklist the current JWT token so it can no longer be used","operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"500":{"description":"Server error"},"401":{"description":"No auth"}},"security":[{"bearerAuth":[]},{"bearerAuth":[]}]}}},"components":{"schemas":{"MovieRatingRequest":{"type":"object","properties":{"rating":{"type":"number","format":"double","description":"${swagger.dto.movie-rating.rating}","example":8.5,"maximum":10.0,"minimum":0.0}}},"JwtUserDetails":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["ADMIN","USER"]}}},"MovieRatingResponse":{"type":"object","properties":{"movieId":{"type":"integer","format":"int64","description":"Unique movie identifier"},"rating":{"type":"number","format":"double","description":"Rating value from 0.0 to 10.0"},"userId":{"type":"integer","format":"int64","description":"Unique user identifier who submitted the rating"}}},"UpdateMovieRequest":{"type":"object","properties":{"nameRussian":{"type":"string","description":"${swagger.dto.create-movie.name-russian}"},"nameNative":{"type":"string","description":"${swagger.dto.create-movie.name-native}"},"picturePath":{"type":"string","description":"${swagger.dto.create-movie.picture-path}"},"countries":{"type":"array","description":"${swagger.dto.create-movie.countries}","items":{"type":"integer","format":"int64"}},"genres":{"type":"array","description":"${swagger.dto.create-movie.genres}","items":{"type":"integer","format":"int64"}}}},"CountryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Unique country identifier"},"name":{"type":"string","description":"Country name"}}},"GenreResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Unique genre identifier"},"name":{"type":"string","description":"Genre name"}}},"MovieDetailResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Unique movie identifier"},"nameRussian":{"type":"string","description":"Movie name in Russian"},"nameNative":{"type":"string","description":"Movie name in original language"},"yearOfRelease":{"type":"integer","format":"int32","description":"Year the movie was released"},"rating":{"type":"number","format":"double","description":"Average movie rating (0.0 - 10.0)"},"price":{"type":"number","format":"double","description":"Movie price in default currency"},"description":{"type":"string","description":"Full description of the movie plot"},"picturePath":{"type":"string","description":"Path to the movie poster image"},"countries":{"type":"array","description":"List of countries associated with the movie","items":{"$ref":"#/components/schemas/CountryResponse"}},"genres":{"type":"array","description":"List of genres associated with the movie","items":{"$ref":"#/components/schemas/GenreResponse"}},"reviews":{"type":"array","description":"List of user reviews for the movie","items":{"$ref":"#/components/schemas/ReviewResponse"}}}},"ReviewResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Unique review identifier"},"comment":{"type":"string","description":"Review text content"},"user":{"$ref":"#/components/schemas/UserResponse","description":"Information about the review author"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Unique user identifier"},"name":{"type":"string","description":"User display name"}}},"CreateReviewRequest":{"type":"object","properties":{"movieId":{"type":"integer","format":"int64","description":"${swagger.dto.create-review.movie-id}"},"text":{"type":"string","description":"${swagger.dto.create-review.text}"}},"required":["movieId","text"]},"TokenResponse":{"type":"object","properties":{"token":{"type":"string","description":"JWT access token"}}},"CreateMovieRequest":{"type":"object","properties":{"nameRussian":{"type":"string","description":"${swagger.dto.create-movie.name-russian}"},"nameNative":{"type":"string","description":"${swagger.dto.create-movie.name-native}"},"yearOfRelease":{"type":"integer","format":"int32","description":"${swagger.dto.create-movie.year-of-release}","minimum":1888},"description":{"type":"string","description":"${swagger.dto.create-movie.description}"},"price":{"type":"number","format":"double","description":"${swagger.dto.create-movie.price}","minimum":0.0},"picturePath":{"type":"string","description":"${swagger.dto.create-movie.picture-path}"},"countries":{"type":"array","description":"${swagger.dto.create-movie.countries}","items":{"type":"integer","format":"int64"}},"genres":{"type":"array","description":"${swagger.dto.create-movie.genres}","items":{"type":"integer","format":"int64"}}},"required":["countries","description","genres","nameNative","nameRussian","picturePath"]},"LoginRequest":{"type":"object","properties":{"email":{"type":"string","description":"${swagger.dto.login.email}"},"password":{"type":"string","description":"${swagger.dto.login.password}"}},"required":["email","password"]},"FindAllMovieRequest":{"type":"object","properties":{"ratingSortDirection":{"type":"string","description":"${swagger.dto.find-all-movie.rating-sort-direction}","enum":["ASC","DESC"],"example":"DESC"},"priceSortDirection":{"type":"string","description":"${swagger.dto.find-all-movie.price-sort-direction}","enum":["ASC","DESC"],"example":"ASC"}}},"MovieResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64","description":"Unique movie identifier"},"nameRussian":{"type":"string","description":"Movie name in Russian"},"nameNative":{"type":"string","description":"Movie name in original language"},"yearOfRelease":{"type":"integer","format":"int32","description":"Year the movie was released"},"rating":{"type":"number","format":"double","description":"Average movie rating (0.0 - 10.0)"},"price":{"type":"number","format":"double","description":"Movie price in default currency"},"picturePath":{"type":"string","description":"Path to the movie poster image"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}